refactor: update build scripts and replace deno scripts with node.js for metadata generation, add CI that actually works
- Changed the `metagen` script from Deno to Node.js in `package.json`.
- Added new build scripts for `libffi` and `metagen` in `package.json`.
- Updated CMakeLists to reflect the change from Deno to Node.js for metadata writing.
- Removed obsolete Deno scripts related to `build_libffi` and `metagen`.
- Deleted GitHub workflows for npm release and pull requests as part of the cleanup.
