+++ NativeScript/libffi/iphonesimulator-universal/include/ffi.h
@@ -1,4 +1,4 @@
-// This file is generated by scripts/build_libffi.ts
+// This file is generated by scripts/build_libffi.js
   // Merged from build_iphonesimulator-arm64/include/ffi.h and build_iphonesimulator-x86_64/include/ffi.h
   
   #if defined(__aarch64__)
