+++ NativeScript/libffi/macosx-universal/include/ffitarget.h
@@ -1,4 +1,4 @@
-// This file is generated by scripts/build_libffi.ts
+// This file is generated by scripts/build_libffi.js
   // Merged from build_macosx-arm64/include/ffitarget.h and build_macosx-x86_64/include/ffitarget.h
   
   #if defined(__aarch64__)
