+++ package.json
@@ -29,7 +29,10 @@
     "update-version": "./update_version.sh",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s",
     "version": "npm run changelog && git add CHANGELOG.md",
-    "metagen": "deno run -A scripts/metagen.ts"
+    "metagen": "node scripts/metagen.js",
+    "build-libffi": "node scripts/build_libffi.js",
+    "build-metagen": "./build_metadata_generator.sh",
+    "build": "./build_nativescript.sh --quickjs"
   },
   "license": "Apache-2.0",
   "devDependencies": {
