+++ scripts/build_libffi.js
@@ -1,0 +1,147 @@
+#!/usr/bin/env node
+
+const { spawnSync } = require("node:child_process");
+const fs = require("node:fs/promises");
+const path = require("node:path");
+
+const archs = ["arm64", "x86_64"];
+
+async function main() {
+  const scriptDir = __dirname;
+  const libffiDir = path.resolve(scriptDir, "..", "libffi");
+  process.chdir(libffiDir);
+
+  process.env.CC = "clang";
+  process.env.CFLAGS = "-w";
+
+  const skipGenerateSource =
+    Boolean(process.env.SKIP_GENERATE_SOURCE) ||
+    process.argv.includes("--skip-generate-source");
+
+  if (!skipGenerateSource) {
+    for (const platform of ["ios", "osx"]) {
+      run("python", [
+        "generate-darwin-source-and-headers.py",
+        `--only-${platform}`,
+      ]);
+    }
+  }
+
+  for (const dir of [
+    "iphoneos-arm64",
+    "iphonesimulator-x86_64",
+    "iphonesimulator-arm64",
+    "macosx-x86_64",
+    "macosx-arm64",
+  ]) {
+    run("make", ["-C", `build_${dir}`, "install"]);
+  }
+
+  await prepareDir("prebuilt/iphoneos-arm64/include");
+  await fs.copyFile(
+    "build_iphoneos-arm64/include/ffi.h",
+    "prebuilt/iphoneos-arm64/include/ffi.h",
+  );
+  await fs.copyFile(
+    "build_iphoneos-arm64/include/ffitarget.h",
+    "prebuilt/iphoneos-arm64/include/ffitarget.h",
+  );
+  await fs.copyFile(
+    "build_iphoneos-arm64/.libs/libffi_convenience.a",
+    "prebuilt/iphoneos-arm64/libffi.a",
+  );
+
+  await prepareDir("prebuilt/macosx-universal/include");
+  await combineHeaders("macosx");
+  run("lipo", [
+    "-create",
+    "-output",
+    "prebuilt/macosx-universal/libffi.a",
+    "build_macosx-x86_64/.libs/libffi_convenience.a",
+    "build_macosx-arm64/.libs/libffi_convenience.a",
+  ]);
+
+  await prepareDir("prebuilt/iphonesimulator-universal/include");
+  run("lipo", [
+    "-create",
+    "-output",
+    "prebuilt/iphonesimulator-universal/libffi.a",
+    "build_iphonesimulator-x86_64/.libs/libffi_convenience.a",
+    "build_iphonesimulator-arm64/.libs/libffi_convenience.a",
+  ]);
+  await combineHeaders("iphonesimulator");
+}
+
+function run(command, args) {
+  console.log(`$ ${command} ${args.join(" ")}`);
+  const result = spawnSync(command, args, {
+    stdio: "inherit",
+    env: process.env,
+  });
+
+  if (result.status !== 0) {
+    throw new Error(
+      `Command failed: ${command} ${args.join(" ")} (exit code ${result.status})`,
+    );
+  }
+}
+
+async function prepareDir(includePath) {
+  const root = path.dirname(includePath);
+  await fs.rm(root, { recursive: true, force: true });
+  await fs.mkdir(includePath, { recursive: true });
+}
+
+async function combineHeaders(target) {
+  const ffi_h_arm64 = await fs.readFile(
+    `build_${target}-${archs[0]}/include/ffi.h`,
+    "utf8",
+  );
+  const ffi_h_x86_64 = await fs.readFile(
+    `build_${target}-${archs[1]}/include/ffi.h`,
+    "utf8",
+  );
+  const ffitarget_h_arm64 = await fs.readFile(
+    `build_${target}-${archs[0]}/include/ffitarget.h`,
+    "utf8",
+  );
+  const ffitarget_h_x86_64 = await fs.readFile(
+    `build_${target}-${archs[1]}/include/ffitarget.h`,
+    "utf8",
+  );
+
+  const ffi_h_universal = `// This file is generated by scripts/build_libffi.js
+  // Merged from build_${target}-${archs[0]}/include/ffi.h and build_${target}-${archs[1]}/include/ffi.h
+  
+  #if defined(__aarch64__)
+  ${ffi_h_arm64}
+  #elif defined(__x86_64__)
+  ${ffi_h_x86_64}
+  #else
+  #error "Unsupported architecture"
+  #endif
+  `;
+
+  const ffitarget_h_universal = `// This file is generated by scripts/build_libffi.js
+  // Merged from build_${target}-${archs[0]}/include/ffitarget.h and build_${target}-${archs[1]}/include/ffitarget.h
+  
+  #if defined(__aarch64__)
+  ${ffitarget_h_arm64}
+  #elif defined(__x86_64__)
+  ${ffitarget_h_x86_64}
+  #else
+  #error "Unsupported architecture"
+  #endif
+  `;
+
+  await fs.writeFile(`prebuilt/${target}-universal/include/ffi.h`, ffi_h_universal);
+  await fs.writeFile(
+    `prebuilt/${target}-universal/include/ffitarget.h`,
+    ffitarget_h_universal,
+  );
+}
+
+main().catch((error) => {
+  console.error(error);
+  process.exit(1);
+});
