--- scripts/build_libffi.ts
@@ -1,135 +1,0 @@
-import { $ } from "./deps.ts";
-
-$.setPrintCommand(true);
-
-// CWD is scripts/
-Deno.chdir(new URL("../libffi", import.meta.url));
-
-Deno.env.set("CC", "clang");
-Deno.env.set("CFLAGS", "-w");
-
-const archs = ["arm64", "x86_64"];
-
-if (
-  !Deno.env.get("SKIP_GENERATE_SOURCE") &&
-  !Deno.args.includes("--skip-generate-source")
-) {
-  for (const platform of ["ios", "osx"]) {
-    await $`python generate-darwin-source-and-headers.py --only-${platform}`;
-  }
-}
-
-for (
-  const dir of [
-    "iphoneos-arm64",
-    "iphonesimulator-x86_64",
-    "iphonesimulator-arm64",
-    "macosx-x86_64",
-    "macosx-arm64",
-  ]
-) {
-  await $`make -C build_${dir} install`;
-}
-
-// iOS
-
-await Deno.remove("prebuilt/iphoneos-arm64", { recursive: true }).catch(
-  () => {},
-);
-await Deno.mkdir("prebuilt/iphoneos-arm64/include", { recursive: true });
-
-await Deno.copyFile(
-  "build_iphoneos-arm64/include/ffi.h",
-  "prebuilt/iphoneos-arm64/include/ffi.h",
-);
-
-await Deno.copyFile(
-  "build_iphoneos-arm64/include/ffitarget.h",
-  "prebuilt/iphoneos-arm64/include/ffitarget.h",
-);
-
-await Deno.copyFile(
-  "build_iphoneos-arm64/.libs/libffi_convenience.a",
-  "prebuilt/iphoneos-arm64/libffi.a",
-);
-
-// macOS
-
-await Deno.remove("prebuilt/macosx-universal", { recursive: true }).catch(
-  () => {},
-);
-await Deno.mkdir("prebuilt/macosx-universal/include", { recursive: true });
-
-await combineHeaders("macosx");
-
-await $`lipo -create -output prebuilt/macosx-universal/libffi.a build_macosx-x86_64/.libs/libffi_convenience.a build_macosx-arm64/.libs/libffi_convenience.a`;
-
-// iOS Simulator
-
-await Deno.remove("prebuilt/iphonesimulator-universal", { recursive: true })
-  .catch(
-    () => {},
-  );
-await Deno.mkdir("prebuilt/iphonesimulator-universal/include", {
-  recursive: true,
-});
-
-await $`lipo -create -output prebuilt/iphonesimulator-universal/libffi.a build_iphonesimulator-x86_64/.libs/libffi_convenience.a build_iphonesimulator-arm64/.libs/libffi_convenience.a`;
-
-await combineHeaders("iphonesimulator");
-
-// Utility
-
-async function combineHeaders(target: string) {
-  const ffi_h_arm64 = await Deno.readTextFile(
-    `build_${target}-${archs[0]}/include/ffi.h`,
-  );
-  const ffi_h_x86_64 = await Deno.readTextFile(
-    `build_${target}-${archs[1]}/include/ffi.h`,
-  );
-  const ffitarget_h_arm64 = await Deno.readTextFile(
-    `build_${target}-${archs[0]}/include/ffitarget.h`,
-  );
-  const ffitarget_h_x86_64 = await Deno.readTextFile(
-    `build_${target}-${archs[1]}/include/ffitarget.h`,
-  );
-
-  const ffi_h_universal = `// This file is generated by scripts/build_libffi.ts
-  // Merged from build_${target}-${
-    archs[0]
-  }/include/ffi.h and build_${target}-${archs[1]}/include/ffi.h
-  
-  #if defined(__aarch64__)
-  ${ffi_h_arm64}
-  #elif defined(__x86_64__)
-  ${ffi_h_x86_64}
-  #else 
-  #error "Unsupported architecture"
-  #endif
-  `;
-
-  const ffitarget_h_universal =
-    `// This file is generated by scripts/build_libffi.ts
-  // Merged from build_${target}-${
-      archs[0]
-    }/include/ffitarget.h and build_${target}-${archs[1]}/include/ffitarget.h
-  
-  #if defined(__aarch64__)
-  ${ffitarget_h_arm64}
-  #elif defined(__x86_64__)
-  ${ffitarget_h_x86_64}
-  #else
-  #error "Unsupported architecture"
-  #endif
-  `;
-
-  await Deno.writeTextFile(
-    `prebuilt/${target}-universal/include/ffi.h`,
-    ffi_h_universal,
-  );
-
-  await Deno.writeTextFile(
-    `prebuilt/${target}-universal/include/ffitarget.h`,
-    ffitarget_h_universal,
-  );
-}
