--- scripts/deps.ts
@@ -1,1 +1,0 @@
-export { default as $ } from "https://deno.land/x/dax@0.32.0/mod.ts";
